added card styling to about page and help page; removed circle decoration from calculator page
