+++ src/Calculator.jsx
@@ -29,27 +29,15 @@ const Calculator = () => {
     <CalculatorNumberButton key={num.val} obj={num} />
   ));
   return (
-    <>
-      <div className="calculator-container">
-        <div className="circle-wrap">
-          <Circle>
-            <p className="circle-caption">
-              the calculator app created by us, for you
-            </p>
-          </Circle>
-        </div>
-
-        <div id="calculatorWrapper">
-          <div id="calculatorBody">
-            <CalculatorDisplay />
-            <div id="buttons">
-              {functionButtons}
-              {numberButtons}
-            </div>
-          </div>
+    <div id="calculatorWrapper">
+      <div id="calculatorBody">
+        <CalculatorDisplay />
+        <div id="buttons">
+          {functionButtons}
+          {numberButtons}
         </div>
       </div>
-    </>
+    </div>
   );
 };
