+++ src/Footer.css
@@ -26,8 +26,7 @@
     font-size:1rem;
     font-weight: 300;
 }
-.footer-list{
-    color:red;
+.footer-list{  
     display: flex;
     justify-content: center;
 }
